backend/requirements.txt
fastapi==0.128.2
uvicorn==0.40.0
pydantic==2.12.5
psycopg2-binary==2.9.9
python-dotenv==1.1.1
bcrypt==5.0.0
PyJWT==2.11.0
langgraph==0.2.20
langchain==0.3.27
langchain-groq==0.3.7
